requirements.txt
manifoldbot>=0.0.1
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.18.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
aiohttp>=3.9.0
python-dateutil>=2.8.2
pydantic>=2.5.0
tenacity>=8.2.3
rich>=13.7.0

